done fetch

+++ app/src/main/java/com/arifinfrds/papblnetworking/model/Wind.java
@@ -9,9 +9,10 @@ public class Wind {
     @SerializedName("speed")
     @Expose
     private Double speed;
+
     @SerializedName("deg")
     @Expose
-    private Integer deg;
+    private Double deg;
 
     public Double getSpeed() {
         return speed;
@@ -21,11 +22,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
